gstvpe: Change NV12 buffer to use one plane instead of two
Starting with TI Kernel 4.9 and mainline Kernel 4.10 the
VPE driver can now handle with single and dual plane
NV12/NV16 buffer.

Realistically NV12 is supposed to be a single plane
format where the Y and UV area are distinct but still
in a single buffer.

Since the various decoder and V4L2 capture engine are
using the single plane representation then keeping that
model when using VPE should be the method used also.

This makes it easier to share DMABUF entities between
the various component and prevent us from having to
implements various hack to make a single plane buffer
look like a dual plane one.

Signed-off-by: Benoit Parrot <bparrot@ti.com>