gstvpe: Keep a copy of the negotiated v4l2 format
When issuing VIDIOC_QBUF ideally the v4l2_buffer structure
being passed should be the exact same one return by the
VIDIOC_QUERYBUF call.

Here since the QUERYBUF info is not kept we need to
at least have all the parameter needed to refilled the
v4l2_buffer structure before passing it to the driver.

Otherwise you'll currently get a kernel WARNING that
you are relying on functionality which is deprecated.

Signed-off-by: Benoit Parrot <bparrot@ti.com>